+++ src/components/Grid/ItemGrid.js
@@ -11,6 +11,9 @@ const style = {
 	noMargin: {
 		margin: "8px 0px"
 	},
+	zeroMargin: {
+		margin: 0
+	},
 	grid: {
 		padding: "0 8px",
 		margin: "8px",
@@ -20,9 +23,9 @@ const style = {
 };
 
 function ItemGrid({ ...props }) {
-	const { classes, children, noPadding, extraClass, noMargin, ...rest } = props;
+	const { classes, children, noPadding, extraClass, noMargin, zeroMargin, ...rest } = props;
 	return (
-		<Grid item {...rest} className={(noPadding ? '' : classes.padding) + " " + (noMargin ? classes.noMargin : classes.margin) + " " + (extraClass ? extraClass : '')}>
+		<Grid item {...rest} className={(zeroMargin ? classes.zeroMargin : '') + " " + (noPadding ? '' : classes.padding) + " " + (noMargin ? classes.noMargin : classes.margin) + " " + (extraClass ? extraClass : '')}>
 			{children}
 		</Grid>
 	);
